moving right works
moving right works, it has some small kinks. Such as being able to make a move and spawn in a new tile when you otherwise shouldnt be able to make a move. but other than that, merges work as intended, and when there are zeros to the right of a number, the nubmer is moved to replace the zeros.