<--- NEW VERSION --->
this is the new version with everything in one script. it is kinda janky and has its kinks, but for the most part it works and does what it is supposed to do, I will need to do a lot of fuckign around to iron oout all the bugs, but it looks pretty good so far. I may have to redo the entire algorithm tho..